requirements.txt
PIL
tensorflow==1.12.0
numpy
matplotlib
tqdm
pyyaml
